+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - [fix] TMCache: cache hits from memory will now update access time on disk
 - [fix] TMDiskCache: set & remove methods now acquire a `UIBackgroundTaskIdentifier`
+- [fix] TMDiskCache: will/didAddObject blocks actually get executed
 - [fix] TMDiskCache: `trimToSize:` now correctly removes objects in order of size
 - [fix] TMMemoryCache: `trimToCost:` now correctly removes objects in order of cost
 - [new] TMDiskCache: added method `trimToSizeByDate:`